+++ tests/Model/EntityModelTest.php
@@ -3,28 +3,20 @@
 namespace Tests\Torr\Rad\Model;
 
 use Doctrine\ORM\EntityManagerInterface;
-use Doctrine\Persistence\ManagerRegistry;
 use PHPUnit\Framework\TestCase;
 use Tests\Torr\Rad\Fixtures\ExampleEntity;
-use Torr\Rad\Entity\Interfaces\EntityInterface;
-use Torr\Rad\Entity\Traits\IdTrait;
-use Torr\Rad\Entity\Traits\TimestampsTrait;
-use Torr\Rad\Model\Model;
+use Torr\Rad\Entity\EntityInterface;
+use Torr\Rad\Entity\ModifiableEntityFieldsTrait;
+use Torr\Rad\Model\EntityModel;
 
-final class ModelTest extends TestCase
+final class EntityModelTest extends TestCase
 {
 	/**
-	 * Tests integration of {@see Model::add()} method
+	 * Tests integration of {@see EntityModel::add()} method
 	 */
 	public function testAdd () : void
 	{
 		$manager = $this->createMock(EntityManagerInterface::class);
-		$registry = $this->createMock(ManagerRegistry::class);
-
-		$registry
-			->method("getManager")
-			->willReturn($manager);
-
 		$entity = new ExampleEntity();
 
 		$manager
@@ -32,24 +24,18 @@ final class ModelTest extends TestCase
 			->method("persist")
 			->with($entity);
 
-		$model = new class ($registry) extends Model {};
+		$model = new class ($manager) extends EntityModel {};
 		$model->add($entity);
 	}
 
 
 	/**
-	 * Tests integration of {@see Model::update()} method
+	 * Tests integration of {@see EntityModel::update()} method
 	 */
 	public function testUpdate () : void
 	{
 		$manager = $this->createMock(EntityManagerInterface::class);
-		$registry = $this->createMock(ManagerRegistry::class);
-
-		$registry
-			->method("getManager")
-			->willReturn($manager);
-
-		$model = new class ($registry) extends Model {};
+		$model = new class ($manager) extends EntityModel {};
 
 		// test with timestamps
 		$entityWithTimestamps = $this->createMock(ExampleEntity::class);
@@ -65,16 +51,11 @@ final class ModelTest extends TestCase
 
 
 	/**
-	 * Tests integration of {@see Model::remove()} method
+	 * Tests integration of {@see EntityModel::remove()} method
 	 */
 	public function testRemove () : void
 	{
 		$manager = $this->createMock(EntityManagerInterface::class);
-		$registry = $this->createMock(ManagerRegistry::class);
-
-		$registry
-			->method("getManager")
-			->willReturn($manager);
 
 		$entity = new ExampleEntity();
 
@@ -83,37 +64,23 @@ final class ModelTest extends TestCase
 			->method("remove")
 			->with($entity);
 
-		$model = new class ($registry) extends Model {};
+		$model = new class ($manager) extends EntityModel {};
 		$model->remove($entity);
 	}
 
 
 	/**
-	 * Tests integration of {@see Model::flush()} method
+	 * Tests integration of {@see EntityModel::flush()} method
 	 */
 	public function testFlush () : void
 	{
 		$manager = $this->createMock(EntityManagerInterface::class);
-		$registry = $this->createMock(ManagerRegistry::class);
-
-		$registry
-			->method("getManager")
-			->willReturn($manager);
 
 		$manager
 			->expects(self::once())
 			->method("flush");
 
-		$model = new class ($registry) extends Model {};
+		$model = new class ($manager) extends EntityModel {};
 		$model->flush();
-	}
-
-	private function createEntity () : EntityInterface
-	{
-		return new class implements EntityInterface
-		{
-			use IdTrait;
-			use TimestampsTrait;
-		};
 	}
 }